feat: 옵션 선택 컴포넌트 하나로 통일
